builds now (and simulation still runs)

+++ protocols/bizcoin/server.go
@@ -77,7 +77,7 @@ func (s *Server) ListenClientTransactions() {
 }
 
 // Instantiate takes blockSize transactions and create the bizcoin instances.
-func (s *Server) Instantiate(node *sda.Node, timeOutMs int) (sda.ProtocolInstance, error) {
+func (s *Server) Instantiate(node *sda.Node, timeOutMs uint64) (sda.ProtocolInstance, error) {
 	// wait until we have enough blocks
 	dbg.Print("waiting enough blocks")
 	currTransactions := s.waitEnoughBlocks()
